pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "iac-agent"
version = "0.1.0"
description = "Local-first Jupyter-style workspace for Infrastructure as Code with conversational AI agent"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.11"
dependencies = [
    # Core Framework
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "websockets>=12.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    
    # AI & LLM Providers
    "openai>=1.7.0",
    "anthropic>=0.8.0",
    "google-generativeai>=0.3.0",
    "azure-ai-ml>=1.12.0",
    "litellm>=1.17.0",
    
    # Agent Framework
    "langgraph>=0.0.30",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-anthropic>=0.0.4",
    
    # Database & State Management
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "alembic>=1.13.0",
    
    # Infrastructure Tools
    "python-terraform>=0.10.1",
    "python-hcl2>=4.3.0",
    "azure-mgmt-resource>=23.0.0",
    "azure-identity>=1.15.0",
    "azure-cli-core>=2.55.0",
    
    # Utilities
    "httpx>=0.26.0",
    "aiofiles>=23.2.1",
    "cryptography>=41.0.0",
    "python-multipart>=0.0.6",
    "jinja2>=3.1.2",
    "rich>=13.7.0",
    
    # Terminal/PTY Support
    "ptyprocess>=0.7.0",
    "pexpect>=4.9.0; sys_platform != 'win32'",
    "pywinpty>=2.0.0; sys_platform == 'win32'",
    
    # Development
    "watchfiles>=0.21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.9.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
    "pre-commit>=3.5.0",
]

[project.scripts]
iac-agent = "iac_agent.cli:main"

[project.urls]
Homepage = "https://github.com/your-username/iac-agent"
Repository = "https://github.com/your-username/iac-agent.git"
Issues = "https://github.com/your-username/iac-agent/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/iac_agent"]

[tool.hatch.envs.default]
dependencies = [
    "coverage[toml]",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
]

[tool.black]
target-version = ['py311']
line-length = 88
skip-string-normalization = true

[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src/iac_agent"]